Lab 1 - Web Proxy/User_Service/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
grpcio==1.66.1
grpcio-tools==1.66.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==5.28.2
setuptools==75.1.0
Werkzeug==3.0.4
psycopg2-binary
PyJWT